requirements.txt
gradio
fastapi
uvicorn
numpy
requests
pymilvus
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests